+++ src/cli/commands/public-cmds/import-cmd.js
@@ -22,6 +22,7 @@ export default class Import extends Command {
     ['f', 'force', 'ignore local changes'],
     ['', 'dist', 'write dist files (when exist) to the configured directory'],
     ['', 'conf', 'write the configuration file (bit.json)'],
+    ['', 'skip-npm-install', 'do not install packages of the imported components'],
     ['', 'ignore-package-json', 'do not generate package.json for the imported component(s)']
   ];
   loader = true;
@@ -40,6 +41,7 @@ export default class Import extends Command {
       force = false,
       dist = false,
       conf = false,
+      skipNpmInstall = false,
       ignorePackageJson = false
     }: {
       tester?: boolean,
@@ -52,6 +54,7 @@ export default class Import extends Command {
       force?: boolean,
       dist?: boolean,
       conf?: boolean,
+      skipNpmInstall?: boolean,
       ignorePackageJson?: boolean
     }
   ): Promise<any> {
@@ -71,6 +74,7 @@ export default class Import extends Command {
       force,
       dist,
       conf,
+      installNpmPackages: !skipNpmInstall,
       withPackageJson: !ignorePackageJson
     }).then(importResults => R.assoc('displayDependencies', displayDependencies, importResults));
   }