+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- `bit import`: install npm packages by default, unless `--skip-npm-install` flag is set
+
 ## [0.12.0-ext.5] - 2017-12-21
 
 - change auto-generated node_modules links to be the same as NPM installation of components (@bit/scope.box.name)